flask-backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
SQLAlchemy==2.0.20
PyJWT==2.8.0
python-dotenv==1.0.0
Werkzeug==2.3.7
Flask-SQLAlchemy==3.0.3
